% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WildfireRisk.R
\docType{package}
\name{BayesFireDanger}
\alias{BayesFireDanger}
\alias{BayesFireDanger-package}
\title{WildfireRisk: Spatial Prediction of Fire Risk}
\description{
WildfireRisk: Spatial Prediction of Fire Risk
}
