% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{calculate_line_risk}
\alias{calculate_line_risk}
\title{Calculate a scan line risk value}
\usage{
calculate_line_risk(tsf_mean, forest_p, distance, is_west, ffdi = 50,
  kbdi = 100)
}
\arguments{
\item{tsf_mean}{Mean value of time since fire for points sampled along
the scan line.}

\item{forest_p}{Proportion of points sampled along the scan line that
were in forest.}

\item{distance}{Line length in metres.}

\item{is_west}{\code{TRUE} if the line has a westerly orientation;
\code{FALSE} otherwise.}

\item{ffdi}{Constant FFDI value to use for calculations (default = 50).}

\item{kbdi}{Constant KBDI value to use for calculations (default = 100).}
}
\description{
Applies the equation of Price \emph{et al.} 2015 to calculate risk
(probability of fire travel along scan line) given line values for mean time
since fire, proportion of forest cover, distance (line length) and whether
the line has a westerly orientation. This function primarily serves as a
helper for the \code{calculate_risk} and \code{treat_blocks} functions.
}
\seealso{
\code{\link{calculate_risk}} \code{\link{treat_blocks}}
}
