% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{calculate_risk}
\alias{calculate_risk}
\title{Calculate risk of wildfire at given locations}
\usage{
calculate_risk(lines, tsf, forest, sample.spacing, dbname = "pointdata.db",
  ffdi = 50, kbdi = 100, tsfmax = 50)
}
\arguments{
\item{lines}{An \code{sf} object (spatial data frame) with scan lines
for each location at which to predict fire risk.}

\item{tsf}{A raster of time since fire values.}

\item{forest}{A raster of forest cover where 0 indicates
non-forest cells and all other values indicate forest cells.}

\item{sample.spacing}{The spacing between sample points along scan lines.}

\item{ffdi}{Constant FFDI value to use for calculations (default = 50).}

\item{kbdi}{Constant KBDI value to use for calculations (default = 100).}

\item{tsfmax}{High value of time since fire to use for calculation of risk
if all scan lines are long unburnt (default = 50).}

\item{point.db}{The path for the SQLite database file in which to cache
sample point data for scan lines. If the specified file already exists
it will be over-written and a warning message issued.}
}
\value{
A \code{risk} object (a type of \code{sf} spatial data frame) containing,
  for each location and scan line, the variables used for the risk calculation
  and the predicted probabilities for both observed and long-unburnt time since
  fire (columns 'pobs' and 'pmax'). These data are also written to the SQLite
  database file. The path/filename of the database with point data is recorded
  as the \code{"dbname"} attribute of the returned object.
}
\description{
This function takes a data frame of scan lines created with \code{make_scan_lines}
and samples time since fire and forest cover values along each line. It then uses
these values to calculate the probability of wildfire travelling along each line.
The calculation is based on the regression equation of Price \emph{et al.} 2015.
The function then calculates a second probability, for comparison, with the time
since fire value for each line set to \code{tsf.max} (a high value representing
long unburnt).
}
\details{
The coordinates and data for points sampled along each scan line are written
to a SQLite database into a table named \code{'pointdata'}. Data for scan
lines, including end-points, location identifiers, values used for the risk
calculation (e.g. mean time since fire) and risk values are also written to
the database into a table named \code{'linedata'}. The path to this file is
recorded as the \code{"dbname"} attribute of the returned \code{risk} object.
If you rename or move the database (e.g. when transferring files to another
system) you should manually update the attribute with the new full path to
the database.
For example: \code{attr(risk, "dbname") <- "c:/foo/bar/pointdata.db"}.
}
\note{
This function assumes that scan lines and raster layers are in
compatible projections with metres as the distance unit. If the input scan lines
have a coordinate reference system defined, it will be set for the output.
}
