% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{dexp_truncated}
\alias{dexp_truncated}
\title{Density function for truncated exponential distribution}
\usage{
dexp_truncated(x, rate, upper)
}
\arguments{
\item{x}{Vector of values.}

\item{rate}{Rate parameter.}
}
\description{
Any values \code{<=0} or \code{>upper} will have density zero.
}
