% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_functions.R
\name{is_line_west}
\alias{is_line_west}
\title{Tests lines for westerly orientation}
\usage{
is_line_west(lines, compass.limits = c(247.5, 292.5))
}
\arguments{
\item{lines}{Either a single \code{"LINESTRING"} object or a list of one or more
objects. If the latter, the list can be a geometry column object (class \code{"sfc"})
from an \code{sf} spatial data frame.}
}
\value{
A logical vector where \code{TRUE} indicates westerly orientation.
}
\description{
Given one or more lines, this function checks if the compass bearing from the start
to the end point of each lies within the range for westerly orientation.
}
