% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_functions.R
\name{line_bearing}
\alias{line_bearing}
\title{Calculates line compass bearings}
\usage{
line_bearing(lines)
}
\arguments{
\item{lines}{Either a single \code{"LINESTRING"} object or a list of one or more
objects. If the latter, the list can be a geometry column object (class \code{"sfc"})
from an \code{sf} spatial data frame.}
}
\value{
Compass bearings of lines as a numeric vector.
}
\description{
Given one or more lines, this function calculates the compass bearing (in
degrees) from the start point to the end point of each.
}
\seealso{
\code{\link[sf]{sfc}}
}
