% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_functions.R
\name{lines_in_blocks}
\alias{lines_in_blocks}
\title{Identify scan lines that intersect polygons}
\usage{
lines_in_blocks(blocks, lines, by = c("block", "line"), strict.crs = FALSE,
  quiet = FALSE)
}
\arguments{
\item{blocks}{Polygons (or multi-polygons). These can be provided
as an \code{sf} data frame or as the path to a file of vector data
(e.g. ESRI shapefile).}

\item{lines}{Scan lines as an \code{sf} object; e.g. as returned by
function \code{make_scan_lines} or \code{calculate_risk}.}

\item{strict.crs}{If TRUE, the blocks and lines objects must have
exactly the same coordinate reference system (CRS) defined. If FALSE,
the function will allow any of the following:
no CRS defined for either object (the function will assume coordinates
can be validly compared between objects);
or CRS defined for only one object (the function will assume it also
applies to the other object);
or a different CRS for each object (the function will transform
block data to have the same CRS as line data).}

\item{quiet}{If TRUE, suppress messages and warnings about coordinate
reference systems of objects.}
}
\description{
Given a set of \emph{N} polygons (e.g. representing management units), this
function returns a list of \emph{N} elements with each element a vector of
integer row indices of the scan lines in the \code{lines} object that
the polygon intersects. A polygon that does not intersect any scan lines
will have a zero-length vector in the list.
}
\seealso{
\code{\link{profile_blocks}}
}
