% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{load_line_risk_table}
\alias{load_line_risk_table}
\title{Load the table of line data and risk values from a database}
\usage{
load_line_risk_table(path)
}
\description{
Reads the table of scan line data and risk values from a SQLite database
(as created by function \code{calculate_risk}) and returns it as a spatial
data frame (a type of \code{sf} object of class \code{risk}).
}
