% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_functions.R
\name{make_scan_lines}
\alias{make_scan_lines}
\title{Generate radial scan lines from one or more locations}
\usage{
make_scan_lines(nlines, locations, lengths, angles = NULL)
}
\arguments{
\item{nlines}{Number of lines to generate for each point.}

\item{locations}{One or more locations represented by one of:
\itemize{
  \item{A vector of X and Y values for a single location;}
  \item{A two-column matrix or data.frame of X and Y values (in that order).}
  \item{An sf object with POINT or MULTIPOINT features. In this case the
    map projection of the input data (if defined) will be set for the output
    scan lines.}
}}

\item{lengths}{Either a single value for uniform line length; or a
vector of \code{n} line lengths; or a function that returns a vector
of lengths.}

\item{angles}{Either a vector of line directions (degrees); or a
function that returns a vector of directions; or \code{NULL}
for uniform distribution of line angles.}
}
\value{
An \code{sf} object (data frame) with columns locationid (integer),
  lineid (integer) and geometry (LINESTRING).
}
\description{
This function generates a set of \code{n} radial scan lines emanating from
each of a given set of locations. Line lengths can be constant; individually
specified; or set by a user-provided function (e.g. to draw them from a
particular distribution). Line angles can be regularly spaced (the default);
individually specified; or set by a user-provided function.
}
\examples{
# Some random point locations in a 20x20km area
Npts <- 100
pts <- matrix(runif(2*Npts, 0, 20000), ncol = 2)

# Generate 80 lines of uniform length 5km from each point.
# Line angles default to regular increments.
lines <- make_scan_lines(80, pts, lengths = 5000)

# Variable line lengths from an exponential distribution
# with mean length 5km (95\% bounds: 125m - 18.4km)
fun <- function(n) rexp(n, 1 / 5000)
lines <- make_scan_lines(80, pts, lengths = fun)

}
