% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_functions.R
\name{mean_exp_truncated}
\alias{mean_exp_truncated}
\title{Mean value of a truncated exponential distribution}
\usage{
mean_exp_truncated(rate, upper)
}
\arguments{
\item{rate}{Vector of one or more values for the rate parameter.}

\item{upper}{Upper (right-hand) limit of the distribution.}
}
\description{
Mean value of a truncated exponential distribution
}
