% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{pointdata_memory}
\alias{pointdata_memory}
\title{Estimate total memory required to load sample point data}
\usage{
pointdata_memory(path, units = c("Mb", "Gb"))
}
\arguments{
\item{path}{Either a character string giving the path and filename
of the SQLite database, or a \code{risk} data frame as returned by
function \code{calculate_risk} which has the path to the database
as an attribute.}

\item{units}{Units in which to report the memory requirement.}
}
\value{
Estimated total memory required.
}
\description{
Sample point data for scan lines is held in a SQLite database file.
This function estimates the total memory required to have all
point data in memory simultaneously. You can use it to check
whether it is safe to run the \code{treat_blocks} function with
\code{in.memory = TRUE} for faster calculations.
}
