% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{profile_blocks}
\alias{profile_blocks}
\title{Profile polygons by scan line risk values}
\usage{
profile_blocks(blocks, risk, intersections = NULL, strict.crs = FALSE,
  quiet = FALSE)
}
\arguments{
\item{blocks}{Polygons (or multi-polygons) to profile. These can be provided
as an \code{sf} data frame or as the path to a file of vector data
(e.g. ESRI shapefile).}

\item{risk}{An \code{sf} data frame of risk values for scan lines as
returned by \code{calculate_risk}.}

\item{intersections}{An optional list of intersection data returned by a
previous call to this function.}

\item{strict.crs}{If TRUE, the blocks and risk objects must have
exactly the same coordinate reference system defined. If FALSE,
the function will make relaxed assumptions about the reference
systems as described for function \code{lines_in_blocks}.}

\item{quiet}{If TRUE, suppress messages and warnings about coordinate
reference systems of objects.}
}
\value{
A list with two elements: blocks and intersections. The blocks
  element is an \code{sf} object with data from the input \code{risk} object
  plus columns giving the number of scan lines intersecting each block;
  number of locations to which the intersected scan lines belong; and
  the mean value of scan line probabilities. The intersections element is
  a list of vectors, one for each block, where vector values are indices
  of intersected scan lines. This can be used with the function on
  subsequent calls to avoid re-doing the intersection between blocks and
  lines.
}
\description{
Profile polygons by scan line risk values
}
\seealso{
\code{\link{lines_in_blocks}}
}
