% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_functions.R
\name{sample_raster}
\alias{sample_raster}
\title{Sample a raster layer at points along scan lines}
\usage{
sample_raster(r, lines, spacing = NULL)
}
\arguments{
\item{r}{The raster layer to be sampled.}

\item{lines}{An \code{sf} object containing scan lines for point locations.}

\item{spacing}{Spacing between adjacent sample points. If \code{NULL}
(the default), this will be set to half the cell width of the raster layer.}
}
\value{
A data frame with columns pointid and lineid (both taken from the input
  \code{lines} object), sampleid and value.
}
\description{
This function takes a raster layer and a samples cell values at regularly
spaced points along a given set of scan lines. The scan lines are provided
as an \code{sf} object as produced by \code{\link{make_scan_lines}}.
}
\note{
If the raster object has multiple layers (ie. a RasterStack or RasterBrick
  object) a warning will be issued and only the first layer will be sampled.
}
\examples{
\dontrun{

# Sample a raster layer of time since fire values
vals <- sample_raster(r.tsf, lines)

# Calculate median time since fire value for the central point
# of each set of scan lines
library(dplyr)

vals \%>\%
  group_by(pointid) \%>\%
  summarize(tsf = median(value, na.rm = TRUE))

}

}
