% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_functions.R
\name{sample_raster_endpoints}
\alias{sample_raster_endpoints}
\title{Sample one or more raster layers at the distal point of each scan line}
\usage{
sample_raster_endpoints(x, lines)
}
\arguments{
\item{x}{Either a single Raster object or a list of Raster objects to sample.
If a named list, the names will be used as the column names for sample values
in the returned data frame.}

\item{lines}{An \code{sf} object containing scan lines for point locations.}
}
\value{
A data frame with columns: locationid, lineid (both taken from the input
  \code{lines} object), x and y ordinates of line endpoint,
  and a column of sample values for each of the input rasters.
}
\description{
This function takes one or more raster layers along with a \code{sf} data frame
of scan lines (as produced by \code{make_scan_lines}) and samples raster values
at the distal point of each scan line.
}
\note{
If a raster object has multiple layers (ie. a RasterStack or RasterBrick
  object) a warning will be issued and only the first layer will be sampled.
}
\examples{
\dontrun{

# Sample a single raster layer of ignition probabilities.
line.igprob <- sample_raster(r.ignition, lines)

}
}
