% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{summarize_location_nbrhood}
\alias{summarize_location_nbrhood}
\title{Summarize raster values in location neighbourhoods}
\usage{
summarize_location_nbrhood(lines, layer, quantiles = c(0.25, 0.75))
}
\arguments{
\item{lines}{An \code{sf} object (spatial data frame) with scan lines
for each location at which to predict fire risk.}

\item{layer}{The raster layer to sample.}

\item{quantiles}{A vector of quantiles (probabilities with values in [0,1])
to calculate in addition to the mean. If \code{NULL} or an empty vector
no quantiles are calculated.}
}
\value{
An \code{sf} data frame of summary statistics for each location.
}
\description{
This function takes a set of scan lines, groups them by location, and summarizes
the values of a given raster layer sampled at the distal end-point of each line.
The resulting summary statistics can be taken as a profile of each location's
neighbourhood. Typical use is with a raster layer of ignition probabilities.
Statistics calculated include the mean and (optionally) specified quantiles.
}
\examples{
\dontrun{

# Sample a raster layer of ignition probabilities to summarize the
# likelihood of ignition in the neighbourhood of each asset location

ignition.locations <- summarize_location_nbrhood(lines, r.ignition.prob)

}

}
\seealso{
\code{\link{summarize_location_risk}}
}
