% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{summarize_location_risk}
\alias{summarize_location_risk}
\title{Summarize scan line risk values by location}
\usage{
summarize_location_risk(line.risk, quantiles = c(0.25, 0.75))
}
\arguments{
\item{line.risk}{Either a \code{risk} object as returned by \code{calculate_risk}
or a character string giving the path and filename of a SQLite database
of risk data.}

\item{quantiles}{A vector of quantiles (probabilities with values in [0,1])
to calculate in addition to the mean. If \code{NULL} or an empty vector
no quantiles are calculated.}
}
\value{
An \code{sf} data frame of summary risk statistics for each location.
}
\description{
This function takes a set of risk values for scan lines,
as returned by \code{calculate_risk}, and summarizes the probability values
over the set of lines for each location by mean and (optionally) specified
quantiles.
}
\examples{
\dontrun{
risk.lines <- calculate_risk(scanlines, tsf, forest, 100)

risk.locations <- summarize_location_risk(risk.lines, quantiles = c(0.1, 0.9))
}

}
\seealso{
\code{\link{calculate_risk}}
}
