% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{treat_blocks}
\alias{treat_blocks}
\title{Simulate prescribed burning of landscape blocks and update risk values}
\usage{
treat_blocks(blocks, risk)
}
\arguments{
\item{blocks}{An \code{sf} object of polygons for which to simulate
prescribed burning.}

\item{risk}{An \code{sf} data frame of scan lines with associated
baseline risk values as returned by function \code{calculate_risk}.}
}
\value{
A data frame of updated scan line risk values with columns:
  locationid, lineid, ptreat.
}
\description{
This function takes one or more blocks (polygons representing management units)
and simulates prescribed burning by setting time since fire within
each block to zero. Data for scan lines passing through the blocks is
then updated and the risk value (probablity of fire traversing the line)
recalculated for each line. The line geometries and their prior risk
values are taken from the \code{risk} object which is an \code{sf}
data frame as returned by \code{calculate_risk}.
}
\details{
Sample point locations and prior data for each scan line are retrieved
from a database file specified by the \code{"dbname"} attribute of
the input \code{risk} object. See \code{calculate_risk} for more details.
}
\examples{
\dontrun{
# Load a polygon shapefile of prescribed burning blocks
blocks <- st_read("burnblocks.shp")

# Simulate treatment for three selected blocks
risk.updated <- treat_blocks(blocks[c(1, 5, 42), ], risk)
}

}
