% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{calculate_risk}
\alias{calculate_risk}
\title{Calculate risk of wildfire at given locations}
\usage{
calculate_risk(lines, tsf, forest, sample.spacing, ffdi = 50, kbdi = 100,
  as.sf = TRUE)
}
\arguments{
\item{lines}{An \code{sf} object (spatial data frame) with scan lines
for each location at which to predict fire risk.}

\item{tsf}{A raster of time since fire values.}

\item{forest}{A raster of forest cover where 0 indicates
non-forest cells and all other values indicate forest cells.}

\item{sample.spacing}{The spacing between sample points along scan lines.}

\item{ffdi}{Constant FFDI value to use for calculations (default = 50).}

\item{kbdi}{Constant KBDI value to use for calculations (default = 100).}

\item{as.sf}{If \code{TRUE} (default), the returned data frame will be an
\code{sf} object where the geometry column has the input scan lines.
If \code{FALSE}, the returned object is a simple data frame with no
geometry column.}
}
\value{
A data frame giving, for each location and scan line, the
  variables used for the risk calculation and the predicted probability.
}
\description{
This function takes a data frame of scan lines created with \code{make_scan_lines}
and samples time since fire and forest cover values along each line. It then uses
these values to calculate the probability of wildfire travelling along each line.
The calculation is based on the regression equation of Price \emph{et al.} 2015.
}
\details{
If the input scan lines have a coordinate reference system, and argument \code{as.sf}
is \code{TRUE} (default), then the same CRS will be set for the output. However,
note that this function does not check that the scan lines and raster layers are in
compatible projections, or that the value for sample.spacing is sensible for
the map units being used. This is up to you!
}
