% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_functions.R
\name{make_scan_lines}
\alias{make_scan_lines}
\title{Generate radial scan lines from one or more locations}
\usage{
make_scan_lines(locations, nlines, lengths, angles = NULL, crs = NULL)
}
\arguments{
\item{locations}{A matrix or data frame with either two columns for X-Y
coordinates; or three columns for location identifier (alpha-numeric)
then X-Y coordinates in that order. If location identifiers are not
provided, these will be set as consecutive integers in the returned
scan lines.}

\item{nlines}{Number of lines to generate for each point.}

\item{lengths}{Either a single value for uniform line length; or a
vector of \code{n} line lengths; or a function that returns a vector
of lengths.}

\item{angles}{Either a vector of line directions (degrees); or a
function that returns a vector of directions; or \code{NULL}
for uniform distribution of line angles.}

\item{crs}{An optional specifier for the coordinate reference system of
the scan lines. This can either be an integer EPSG code, a character
string in proj4 format, or a spatial object from which a map projection
can be taken (e.g. a Raster object or an 'sf' spatial data frame).}
}
\value{
An \code{sf} object (data frame) with columns locationid (integer),
  lineid (integer) and geometry (LINESTRING).
}
\description{
This function generates a set of \code{n} radial scan lines emanating from
each of a given set of locations. Line lengths can be constant; individually
specified; or set by a user-provided function (e.g. to draw them from a
particular distribution). Line angles can be regularly spaced (the default);
individually specified; or set by a user-provided function.
}
\examples{
# Some random point locations in a 20x20km area
Npts <- 10
pts <- matrix(runif(2*Npts, 0, 20000), ncol = 2)

# Add identifying labels for locations
pts <- data.frame(id = LETTERS[1:Npts], pts)
colnames(pts)[2:3] <- c("x", "y")

# Generate 80 lines of uniform length 5km from each point.
# Line angles default to regular increments.
# Set the coordinate reference system to MGA Zone 55
# using its EPSG code 28355
#
lines <- make_scan_lines(80, pts, lengths = 5000, crs = 28355)

# Variable line lengths from a truncated exponential distribution
# shifted so that minimum line length is 500m, with upper limit set
# so that maximum line length is 20km
#
fun <- function(n) 500 + rexp_truncated(n, 1/5000, 19500)
lines <- make_scan_lines(80, pts, lengths = fun)

}
