% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_functions.R
\name{summarize_risk}
\alias{summarize_risk}
\title{Summarize scan line risk values}
\usage{
summarize_risk(risk)
}
\arguments{
\item{risk}{An \code{sf} data frame of risk values for scan lines as
returned by \code{calculate_risk}.}
}
\value{
An \code{sf} data frame of summary risk statistics for each location.
}
\description{
This function takes an \code{sf} data frame of risk values for scan lines,
as returned by \code{calculate_risk}, and summarizes the probability values
over the set of lines for each location by: mean; median; lower quartile and
upper quartile.
}
\seealso{
\code{\link{calculate_risk}}
}
